+++ src/components/graphview/GraphViewPanel.tsx
@@ -69,6 +69,8 @@ export interface GraphViewPanelProps {
   onRemoveOverlay?: (id: string) => void;
   alignOverlays?: boolean;
   onToggleAlignOverlays?: () => void;
+  overlaysVisible?: boolean;
+  onToggleOverlaysVisible?: () => void;
 }
 
 export function GraphViewPanel(props: GraphViewPanelProps) {
@@ -159,6 +161,8 @@ export function GraphViewPanel(props: GraphViewPanelProps) {
                 onRemoveOverlay={props.onRemoveOverlay}
                 alignOverlays={props.alignOverlays}
                 onToggleAlignOverlays={props.onToggleAlignOverlays}
+                overlaysVisible={props.overlaysVisible}
+                onToggleOverlaysVisible={props.onToggleOverlaysVisible}
               />
             </ResizablePanel>
           </ResizablePanelGroup>